schema: Fix wrong type of CLTV in route
We were using `number` instead of `u32` which caused it to be
serialized to string with a decimal, which then would be rejected by
lightningd.

Changelog-Fixed: proto: Fixed a wrong number type being used in routes

+++ cln-rpc/src/model.rs
@@ -1270,7 +1270,7 @@ pub mod requests {
 	    pub amount_msat: Amount,
 	    pub riskfactor: u64,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub cltv: Option<f64>,
+	    pub cltv: Option<u32>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub fromid: Option<PublicKey>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
